+++ spec/features/projects/in_progress/users_can_view_a_list_of_all_projects_spec.rb
@@ -9,7 +9,7 @@ RSpec.feature "Viewing all in-progress projects" do
 
   context "when there are no projects" do
     scenario "they can see a helpful message" do
-      visit all_in_progress_projects_path
+      visit all_all_in_progress_projects_path
 
       expect(page).to have_content(I18n.t("project.table.in_progress.empty"))
     end
@@ -20,7 +20,7 @@ RSpec.feature "Viewing all in-progress projects" do
       completed_project = create(:conversion_project, :completed, urn: 121583, completed_at: Date.yesterday)
       in_progress_project = create(:conversion_project, urn: 115652)
 
-      visit all_in_progress_projects_path
+      visit all_all_in_progress_projects_path
 
       expect(page).to have_content(I18n.t("project.all.in_progress.title"))
 
@@ -35,7 +35,7 @@ RSpec.feature "Viewing all in-progress projects" do
       last_project = create(:conversion_project, conversion_date: Date.parse("2023-12-1"), urn: 165432)
       first_project = create(:conversion_project, conversion_date: Date.parse("2023-1-1"), urn: 123456)
 
-      visit all_in_progress_projects_path
+      visit all_all_in_progress_projects_path
 
       within "tbody tr:first-child" do
         expect(page).to have_content(first_project.urn)
@@ -50,7 +50,7 @@ RSpec.feature "Viewing all in-progress projects" do
         create(:conversion_project, region: :london)
       end
 
-      visit all_in_progress_projects_path
+      visit all_all_in_progress_projects_path
 
       expect(page).to have_css(".govuk-pagination")
     end
@@ -58,7 +58,7 @@ RSpec.feature "Viewing all in-progress projects" do
     scenario "the projects show if they are 'Form a MAT' or not" do
       _project = create(:conversion_project, :form_a_mat, urn: 115652)
 
-      visit all_in_progress_projects_path
+      visit all_all_in_progress_projects_path
       expect(page).to have_content("Form a MAT project?")
       expect(page).to have_content("Yes")
     end
@@ -67,7 +67,7 @@ RSpec.feature "Viewing all in-progress projects" do
       deleted_project = create(:conversion_project, :deleted, urn: 121583)
       in_progress_project = create(:conversion_project, urn: 115652)
 
-      visit all_in_progress_projects_path
+      visit all_all_in_progress_projects_path
 
       expect(page).to_not have_content(deleted_project.urn.to_s)
       expect(page).to have_content(in_progress_project.urn.to_s)